Consolidate metrics by whether emitted by metrics collector
 * all gauge metrics that require state are moved to syncer_metrics.go
   are published by the metrics collector

 * all metrics that do not require state and not emitted by the metrics
   collector
