--- pkg/neg/metrics/endpoint_metrics.go
@@ -1,43 +1,0 @@
-/*
-Copyright 2023 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package metrics
-
-import (
-	"github.com/prometheus/client_golang/prometheus"
-)
-
-var (
-	// syncerEndpointState tracks the count of endpoints in different states
-	syncerEndpointState = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Subsystem: negControllerSubsystem,
-			Name:      "syncer_endpoint_state",
-			Help:      "Current count of endpoints in each state",
-		},
-		[]string{"state"},
-	)
-
-	// syncerEndpointSliceState tracks the count of endpoint slices in different states
-	syncerEndpointSliceState = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Subsystem: negControllerSubsystem,
-			Name:      "syncer_endpoint_slice_state",
-			Help:      "Current count of endpoint slices in each state",
-		},
-		[]string{"state"},
-	)
-)
