+++ pkg/neg/metrics/neg_metrics_collector.go
@@ -86,10 +86,15 @@ func FakeSyncerMetrics() *SyncerMetrics {
 
 // RegisterSyncerMetrics registers syncer related metrics
 func RegisterSyncerMetrics() {
-	prometheus.MustRegister(syncerSyncResult)
 	prometheus.MustRegister(syncerState)
 	prometheus.MustRegister(syncerEndpointState)
 	prometheus.MustRegister(syncerEndpointSliceState)
+	prometheus.MustRegister(NumberOfEndpoints)
+	prometheus.MustRegister(DualStackMigrationFinishedDurations)
+	prometheus.MustRegister(DualStackMigrationLongestUnfinishedDuration)
+	prometheus.MustRegister(DualStackMigrationServiceCount)
+	prometheus.MustRegister(SyncerCountByEndpointType)
+	prometheus.MustRegister(syncerSyncResult)
 }
 
 func (sm *SyncerMetrics) Run(stopCh <-chan struct{}) {
@@ -372,4 +377,24 @@ func (sm *SyncerMetrics) computeDualStackMigrationCounts() (map[string]int, int,
 		}
 	}
 	return syncerCountByEndpointType, migrationEndpointCount, migrationServices.Len()
+}
+
+func PublishSyncerStateMetrics(stateCount *syncerStateCount) {
+	syncerState.WithLabelValues(EPCountsDiffer).Set(float64(stateCount.epCountsDiffer))
+	syncerState.WithLabelValues(EPNodeMissing).Set(float64(stateCount.epNodeMissing))
+	syncerState.WithLabelValues(EPNodeNotFound).Set(float64(stateCount.epNodeNotFound))
+	syncerState.WithLabelValues(EPPodMissing).Set(float64(stateCount.epPodMissing))
+	syncerState.WithLabelValues(EPPodNotFound).Set(float64(stateCount.epPodNotFound))
+	syncerState.WithLabelValues(EPPodTypeAssertionFailed).Set(float64(stateCount.epPodTypeAssertionFailed))
+	syncerState.WithLabelValues(EPZoneMissing).Set(float64(stateCount.epZoneMissing))
+	syncerState.WithLabelValues(EPSEndpointCountZero).Set(float64(stateCount.epsEndpointCountZero))
+	syncerState.WithLabelValues(EPCalculationCountZero).Set(float64(stateCount.epCalculationCountZero))
+	syncerState.WithLabelValues(InvalidAPIResponse).Set(float64(stateCount.invalidAPIResponse))
+	syncerState.WithLabelValues(InvalidEPAttach).Set(float64(stateCount.invalidEPAttach))
+	syncerState.WithLabelValues(InvalidEPDetach).Set(float64(stateCount.invalidEPDetach))
+	syncerState.WithLabelValues(NegNotFound).Set(float64(stateCount.negNotFound))
+	syncerState.WithLabelValues(CurrentNegEPNotFound).Set(float64(stateCount.currentNegEPNotFound))
+	syncerState.WithLabelValues(EPSNotFound).Set(float64(stateCount.epsNotFound))
+	syncerState.WithLabelValues(OtherError).Set(float64(stateCount.otherError))
+	syncerState.WithLabelValues(Success).Set(float64(stateCount.success))
 }
